Updated recipe endpoints

+++ backend/services/recipe.ts
@@ -1,48 +1,64 @@
-import { Recipe, RecipeSmall } from '../models/recipe';
+import {
+    updateNamespaceExportDeclaration
+} from 'typescript';
+import {
+    Recipe,
+    RecipeSmall
+} from '../models/recipe';
 import * as recipeDAO from '../storage/recipe';
 import * as tagDAO from '../storage/tag';
-import { generateUUID } from '../util/uuid';
+import * as userDAO from '../storage/user';
+import * as categoryDAO from '../storage/category';
 
-export function createRecipe(recipe: Recipe): Promise<Recipe> {
+import {
+    generateUUID
+} from '../util/uuid';
+
+export function createRecipe(recipe: Recipe): Promise < Recipe > {
     recipe.id = generateUUID();
     recipe.createdAt = new Date();
 
     //TODO Check if category exists
 
-    return new Promise<Recipe>((resolve, reject) => {
-        recipeDAO.createRecipe(recipe).then(recipe => {
-            //Create all ingredients
-            recipe.ingredients.forEach(ingredient => {
-                ingredient.id = generateUUID();
-                ingredient.recipe_id = recipe.id;
-                //Create ingredient
-                recipeDAO.createIngredient(ingredient).catch(err => {
-                    reject(err);
-                });
-            });
-
-            //Create all tags and add them to recipe
-            recipe.tags.forEach(tag => {
-                //Create tag if not exists
-                tagDAO.createTag(tag).then().catch(err => {
-                    //Dont catch error (tag already exists)
-                }).finally(() => {
-                    //Add tag to recipe
-                    tagDAO.addRecipeTag(recipe.id, tag.name).catch(err => {
-                        reject(err);
+    return new Promise < Recipe > ((resolve, reject) => {
+        //Check if category exits
+        categoryDAO.getCategory(recipe.category_id).then((category) => {
+            if (category) {
+                recipeDAO.createRecipe(recipe).then(recipe => {
+                    //Create all ingredients
+                    recipe.ingredients.forEach(ingredient => {
+                        ingredient.id = generateUUID();
+                        ingredient.recipe_id = recipe.id;
+                        //Create ingredient
+                        recipeDAO.createIngredient(ingredient).catch(err => {
+                            reject(err);
+                        });
                     });
+                    //Create all tags and add them to recipe
+                    recipe.tags.forEach(tag => {
+                        //Create tag if not exists
+                        tagDAO.createTag(tag).then().catch(err => {
+                            //Dont catch error (tag already exists)
+                        }).finally(() => {
+                            //Add tag to recipe
+                            tagDAO.addRecipeTag(recipe.id, tag.name).catch(err => {
+                                reject(err);
+                            });
+                        });
+                    });
+                    resolve(recipe);
+                }).catch(error => {
+                    reject(error);
                 });
-            });
-
-            resolve(recipe);
-        }).catch(error => {
-            reject(error);
+            }
+        }).catch((err) => {
+            reject(new Error('Category does not exist'));
         });
     });
-}   
+}
 
-export function getRecipe(recipeID: string): Promise<Recipe> {
-    return new Promise<Recipe>((resolve, reject) => {
+export function getRecipe(recipeID: string): Promise < Recipe > {
+    return new Promise < Recipe > ((resolve, reject) => {
         recipeDAO.getRecipe(recipeID).then(recipe => {
             //Get all ingredients
             recipeDAO.getIngredients(recipe.id).then(ingredients => {
@@ -67,8 +83,8 @@ export function getRecipe(recipeID: string): Promise<Recipe> {
     });
 }
 
-export function getAllRecipes(): Promise<RecipeSmall[]> {
-    return new Promise<RecipeSmall[]>((resolve, reject) => {
+export function getAllRecipes(): Promise < RecipeSmall[] > {
+    return new Promise < RecipeSmall[] > ((resolve, reject) => {
         recipeDAO.getRecipes().then(recipes => {
             let allRecipes: RecipeSmall[] = [];
             recipes.forEach(recipe => {
@@ -91,26 +107,38 @@ export function getAllRecipes(): Promise<RecipeSmall[]> {
     });
 }
 
-export function deleteRecipe(recipeID: string): Promise<void> {
-    return new Promise<void>((resolve, reject) => {
-        //Delete all ingredients
-        recipeDAO.getIngredients(recipeID).then(ingredients => {
-            ingredients.forEach(ingredient => {
-                recipeDAO.deleteIngredient(ingredient.id).catch(err => {
-                    reject(err);
-                });
+export function deleteRecipe(recipeID: string, userID: string): Promise < void > {
+    return new Promise < void > ((resolve, reject) => {
+        //Check if user is owner of recipe
+        recipeDAO.getRecipe(recipeID).then(recipe => {
+            userDAO.getUserById(userID).then(user => {
+                if (user.role === "admin" || user.id === userID) {
+                    //Delete all ingredients
+                    recipeDAO.getIngredients(recipeID).then(ingredients => {
+                        ingredients.forEach(ingredient => {
+                            recipeDAO.deleteIngredient(ingredient.id).catch(err => {
+                                reject(err);
+                            });
+                            //TODO Delete all ratings
+                            //TODO Delete all comments
+                            //Delete Recipe
+                            recipeDAO.deleteRecipe(recipeID).then(() => {
+                                resolve();
+                            }).catch(err => {
+                                reject(err);
+                            });
+                        });
+                    }).catch(error => {
+                        reject(error);
+                    });
+                } else {
+                    reject("User is not owner of recipe");
+                }
+            }).catch(err => {
+                reject(err);
             });
         }).catch(error => {
-            reject(error);
-        });
-
-        //TODO Delete all ratings
-        //TODO Delete all comments
-
-        //Delete Recipe
-        recipeDAO.deleteRecipe(recipeID).then(() => {
-            resolve();
-        }).catch(error => {
+            console.log(error);
             reject(error);
         });
     });
